moved HTML table generation to separate classes - part 1 complete: ConnectomeToHTML
